Message_Parsing/Backend/app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.8.0
python-dotenv==1.0.1
langchain-google-genai==1.0.8
certifi==2024.8.30
python-multipart